refactor(import-export): consolidate to one home, delete dead pipeline
The importExport/ directory shipped a second, never-used import/export pipeline
(index.ts, importService.ts) built on an incompatible CleanExportData v3 format
with zero callers — including a duplicate `importData`. Delete it. The flat
`importExport.ts` is the canonical public API (contentPacks, PackImportDialog,
CustomTileDialog, tests all resolve to it). Keep only the two live UI helpers
the app reaches by subpath: getExportableGroupStats, batchFetchAllGroups.
Public import paths unchanged — no caller edits.
